[0.13.0][Bugfix] Fixed an accuracy problem of sp with eagle3
